+++ src/main/java/com/shortkki/api/recipeImport/service/RecipeImportAsyncService.java
@@ -1,0 +1,205 @@
+package com.shortkki.api.recipeImport.service;
+
+import com.shortkki.api.ingredient.entity.Ingredient;
+import com.shortkki.api.ingredient.repository.IngredientRepository;
+import com.shortkki.api.member.entity.Member;
+import com.shortkki.api.member.repository.MemberRepository;
+import com.shortkki.api.recipe.constant.CuisineType;
+import com.shortkki.api.recipe.constant.Difficulty;
+import com.shortkki.api.recipe.constant.MealType;
+import com.shortkki.api.recipe.entity.Recipe;
+import com.shortkki.api.recipe.entity.RecipeIngredient;
+import com.shortkki.api.recipe.entity.RecipeStep;
+import com.shortkki.api.recipe.repository.RecipeIngredientRepository;
+import com.shortkki.api.recipe.repository.RecipeRepository;
+import com.shortkki.api.recipe.repository.RecipeStepRepository;
+import com.shortkki.api.recipeBook.entity.RecipeBook;
+import com.shortkki.api.recipeBook.service.RecipeBookQueryService;
+import com.shortkki.api.recipeBook.service.RecipeBookService;
+import com.shortkki.api.recipeImport.dto.RecipeParseResult;
+import com.shortkki.api.recipeImport.dto.RecipeParseResult.IngredientParseResult;
+import com.shortkki.api.recipeImport.dto.RecipeParseResult.StepParseResult;
+import com.shortkki.api.recipeImport.service.parser.AiRecipeParserService;
+import com.shortkki.api.source.domain.SourceContent;
+import com.shortkki.api.source.domain.ImportStatus;
+import com.shortkki.api.source.domain.SourceImportHistory;
+import com.shortkki.api.source.repository.SourceContentRepository;
+import com.shortkki.api.source.repository.SourceImportHistoryRepository;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.scheduling.annotation.Async;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.PlatformTransactionManager;
+import org.springframework.transaction.support.TransactionTemplate;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class RecipeImportAsyncService {
+
+    private final SourceContentRepository sourceContentRepository;
+    private final MemberRepository memberRepository;
+    private final RecipeRepository recipeRepository;
+    private final RecipeBookService recipeBookService;
+    private final RecipeBookQueryService recipeBookQueryService;
+    private final SourceImportHistoryRepository sourceImportHistoryRepository;
+    private final AiRecipeParserService aiRecipeParserService;
+    private final IngredientRepository ingredientRepository;
+    private final RecipeIngredientRepository recipeIngredientRepository;
+    private final RecipeStepRepository recipeStepRepository;
+    private final PlatformTransactionManager transactionManager;
+
+    @Async
+    public void processImport(
+            Long memberId,
+            Long sourceContentId,
+            Long historyId,
+            String sourceUrl
+    ) {
+        SourceImportHistory history = sourceImportHistoryRepository.findById(historyId)
+                .orElse(null);
+        if (history == null) {
+            log.warn("Source import history not found. historyId={}", historyId);
+            return;
+        }
+
+        Member member = memberRepository.findById(memberId)
+                .orElse(null);
+        SourceContent sourceContent = sourceContentRepository.findById(sourceContentId)
+                .orElse(null);
+
+        if (member == null || sourceContent == null) {
+            history.fail("Member or source content not found.");
+            sourceImportHistoryRepository.save(history);
+            return;
+        }
+
+        RecipeParseResult parseResult;
+        try {
+            history.updateStatus(ImportStatus.PARSING);
+            sourceImportHistoryRepository.save(history);
+            log.info("AI 레시피 파싱 시작: {}", sourceUrl);
+            parseResult = aiRecipeParserService.parseRecipeFromUrl(sourceUrl);
+            log.info("AI 파싱 완료 - 제목: {}, 재료: {}개, 순서: {}개",
+                    parseResult.title(),
+                    parseResult.ingredients().size(),
+                    parseResult.steps().size());
+        } catch (Exception e) {
+            history.fail("AI parsing failed: " + e.getMessage());
+            sourceImportHistoryRepository.save(history);
+            return;
+        }
+
+        Recipe saved = new TransactionTemplate(transactionManager).execute(status -> {
+            Recipe recipe = saveRecipe(member, sourceContent, parseResult);
+            saveIngredients(recipe, parseResult.ingredients());
+            saveSteps(recipe, parseResult.steps());
+            return recipe;
+        });
+
+        if (saved == null) {
+            history.fail("Recipe transaction rolled back.");
+            sourceImportHistoryRepository.save(history);
+            return;
+        }
+
+        // TODO: 태그 저장 로직 추가
+
+        history.complete("Recipe created: " + saved.getId());
+        sourceImportHistoryRepository.save(history);
+
+        // TODO: 파싱 완료 알림 처리
+        addToDefaultRecipeBook(memberId, saved.getId());
+    }
+
+    private Recipe saveRecipe(Member member, SourceContent sourceContent, RecipeParseResult parseResult) {
+        Recipe recipe = Recipe.createImported(
+                member,
+                parseResult.title() != null ? parseResult.title() : sourceContent.getTitle(),
+                parseResult.description(),
+                parseResult.servingSize(),
+                parseResult.cookingTime(),
+                parseCuisineType(parseResult.cuisineType()),
+                parseMealType(parseResult.mealType()),
+                parseDifficulty(parseResult.difficulty()),
+                sourceContent);
+        return recipeRepository.save(recipe);
+    }
+
+    private void saveIngredients(Recipe recipe, List<IngredientParseResult> ingredients) {
+        for (IngredientParseResult ing : ingredients) {
+            Ingredient ingredient = ingredientRepository.findByName(ing.name())
+                    .orElseGet(() -> ingredientRepository.save(
+                            Ingredient.create(ing.name(), ing.amount())));
+            Integer amount = parseAmount(ing.amount());
+            RecipeIngredient recipeIngredient = RecipeIngredient.create(
+                    ingredient, recipe, amount);
+            recipeIngredientRepository.save(recipeIngredient);
+        }
+    }
+
+    private Integer parseAmount(String amountStr) {
+        if (amountStr == null || amountStr.isBlank()) {
+            return null;
+        }
+        try {
+            String numberOnly = amountStr.replaceAll("[^0-9]", "");
+            return numberOnly.isEmpty() ? null : Integer.parseInt(numberOnly);
+        } catch (NumberFormatException e) {
+            return null;
+        }
+    }
+
+    private void saveSteps(Recipe recipe, List<StepParseResult> steps) {
+        for (StepParseResult step : steps) {
+            RecipeStep recipeStep = RecipeStep.create(
+                    recipe, step.stepNumber(), step.description());
+            recipeStepRepository.save(recipeStep);
+        }
+    }
+
+    private void addToDefaultRecipeBook(Long memberId, Long recipeId) {
+        recipeBookQueryService.findAllByMemberId(memberId).stream()
+                .filter(RecipeBook::getIsDefault)
+                .findFirst()
+                .ifPresent(defaultBook -> recipeBookService.addRecipeInternal(
+                        memberId, defaultBook.getId(), recipeId));
+    }
+
+    private CuisineType parseCuisineType(String value) {
+        if (value == null) {
+            return null;
+        }
+        try {
+            return CuisineType.valueOf(value);
+        } catch (IllegalArgumentException e) {
+            log.warn("Unknown cuisineType: {}", value);
+            return null;
+        }
+    }
+
+    private MealType parseMealType(String value) {
+        if (value == null) {
+            return null;
+        }
+        try {
+            return MealType.valueOf(value);
+        } catch (IllegalArgumentException e) {
+            log.warn("Unknown mealType: {}", value);
+            return null;
+        }
+    }
+
+    private Difficulty parseDifficulty(String value) {
+        if (value == null) {
+            return null;
+        }
+        try {
+            return Difficulty.valueOf(value);
+        } catch (IllegalArgumentException e) {
+            log.warn("Unknown difficulty: {}", value);
+            return null;
+        }
+    }
+}
