+++ src/main/java/com/shortkki/api/recipeImport/service/RecipeImportService.java
@@ -1,27 +1,10 @@
 package com.shortkki.api.recipeImport.service;
 
-import com.shortkki.api.ingredient.entity.Ingredient;
-import com.shortkki.api.ingredient.repository.IngredientRepository;
 import com.shortkki.api.member.entity.Member;
 import com.shortkki.api.member.repository.MemberRepository;
-import com.shortkki.api.recipe.constant.CuisineType;
-import com.shortkki.api.recipe.constant.Difficulty;
-import com.shortkki.api.recipe.constant.MealType;
-import com.shortkki.api.recipe.entity.Recipe;
-import com.shortkki.api.recipe.entity.RecipeIngredient;
-import com.shortkki.api.recipe.entity.RecipeStep;
-import com.shortkki.api.recipe.repository.RecipeIngredientRepository;
-import com.shortkki.api.recipe.repository.RecipeRepository;
-import com.shortkki.api.recipe.repository.RecipeStepRepository;
-import com.shortkki.api.recipeBook.entity.RecipeBook;
-import com.shortkki.api.recipeBook.service.RecipeBookQueryService;
-import com.shortkki.api.recipeBook.service.RecipeBookService;
 import com.shortkki.api.recipeImport.dto.RecipeImportRequest;
 import com.shortkki.api.recipeImport.dto.RecipeImportResponse;
-import com.shortkki.api.recipeImport.dto.RecipeParseResult;
-import com.shortkki.api.recipeImport.dto.RecipeParseResult.IngredientParseResult;
-import com.shortkki.api.recipeImport.dto.RecipeParseResult.StepParseResult;
-import com.shortkki.api.recipeImport.service.parser.AiRecipeParserService;
+import com.shortkki.api.recipeImport.dto.RecipeImportStatusResponse;
 import com.shortkki.api.source.domain.SourceContent;
 import com.shortkki.api.source.domain.SourceImportHistory;
 import com.shortkki.api.source.domain.SourcePlatform;
@@ -30,19 +13,13 @@ import com.shortkki.api.source.repository.SourceImportHistoryRepository;
 import com.shortkki.api.source.service.SourceContentService;
 import com.shortkki.api.source.support.ExternalKeyExtractorRegistry;
 import com.shortkki.global.error.ErrorCode;
+import com.shortkki.global.error.exception.AccessDeniedException;
 import com.shortkki.global.error.exception.BadRequestException;
 import com.shortkki.global.error.exception.BusinessException;
 import com.shortkki.global.error.exception.NotFoundException;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.support.TransactionTemplate;
 
-import java.util.List;
-
-@Slf4j
 @Service
 @RequiredArgsConstructor
 public class RecipeImportService {
@@ -50,16 +27,9 @@ public class RecipeImportService {
     private final SourceContentService sourceContentService;
     private final SourceContentRepository sourceContentRepository;
     private final ExternalKeyExtractorRegistry extractorRegistry;
-    private final RecipeRepository recipeRepository;
     private final MemberRepository memberRepository;
-    private final RecipeBookService recipeBookService;
-    private final RecipeBookQueryService recipeBookQueryService;
     private final SourceImportHistoryRepository sourceImportHistoryRepository;
-    private final AiRecipeParserService aiRecipeParserService;
-    private final IngredientRepository ingredientRepository;
-    private final RecipeIngredientRepository recipeIngredientRepository;
-    private final RecipeStepRepository recipeStepRepository;
-    private final PlatformTransactionManager transactionManager;
+    private final RecipeImportAsyncService recipeImportAsyncService;
 
     public RecipeImportResponse importFromUrl(Long memberId, RecipeImportRequest request) {
         Member member = findMemberById(memberId);
@@ -70,135 +40,35 @@ public class RecipeImportService {
         SourceContent sourceContent = sourceContentService.resolveSourceContent(sourceUrl);
 
         SourceImportHistory history = SourceImportHistory.create(
+                member.getId(),
                 sourceContent.getId(),
                 sourceUrl,
                 sourceContent.getPlatform());
         sourceImportHistoryRepository.save(history);
 
-        RecipeParseResult parseResult;
-        try {
-            log.info("AI 레시피 파싱 시작: {}", sourceUrl);
-            parseResult = aiRecipeParserService.parseRecipeFromUrl(sourceUrl);
-            log.info("AI 파싱 완료 - 제목: {}, 재료: {}개, 순서: {}개",
-                    parseResult.title(),
-                    parseResult.ingredients().size(),
-                    parseResult.steps().size());
-        } catch (Exception e) {
-            history.fail("AI parsing failed: " + e.getMessage());
-            sourceImportHistoryRepository.save(history);
-            throw e;
-        }
+        recipeImportAsyncService.processImport(
+                member.getId(),
+                sourceContent.getId(),
+                history.getId(),
+                sourceUrl);
 
-        Recipe saved = new TransactionTemplate(transactionManager).execute(status -> {
-            Recipe recipe = saveRecipe(member, sourceContent, parseResult);
-            saveIngredients(recipe, parseResult.ingredients());
-            saveSteps(recipe, parseResult.steps());
-            return recipe;
-        });
-
-        // TODO: 태그 저장 로직 추가
-
-        history.complete("Recipe created: " + saved.getId());
-        sourceImportHistoryRepository.save(history);
-
-        addToDefaultRecipeBook(memberId, saved.getId());
-
-        return RecipeImportResponse.success(saved.getId(), saved.getTitle(), sourceUrl);
+        return RecipeImportResponse.accepted(history.getId(), sourceUrl);
     }
 
-    @Transactional
-    private Recipe saveRecipe(Member member, SourceContent sourceContent, RecipeParseResult parseResult) {
-        Recipe recipe = Recipe.createImported(
-                member,
-                parseResult.title() != null ? parseResult.title() : sourceContent.getTitle(),
-                parseResult.description(),
-                parseResult.servingSize(),
-                parseResult.cookingTime(),
-                parseCuisineType(parseResult.cuisineType()),
-                parseMealType(parseResult.mealType()),
-                parseDifficulty(parseResult.difficulty()),
-                sourceContent);
-        return recipeRepository.save(recipe);
-    }
+    public RecipeImportStatusResponse getStatus(Long memberId, Long historyId) {
+        SourceImportHistory history = sourceImportHistoryRepository.findById(historyId)
+                .orElseThrow(() -> new NotFoundException(ErrorCode.NOT_FOUND_ERROR));
 
-    @Transactional
-    private void saveIngredients(Recipe recipe, List<IngredientParseResult> ingredients) {
-        for (IngredientParseResult ing : ingredients) {
-            Ingredient ingredient = ingredientRepository.findByName(ing.name())
-                    .orElseGet(() -> ingredientRepository.save(
-                            Ingredient.create(ing.name(), ing.amount())));
-            Integer amount = parseAmount(ing.amount());
-            RecipeIngredient recipeIngredient = RecipeIngredient.create(
-                    ingredient, recipe, amount);
-            recipeIngredientRepository.save(recipeIngredient);
+        if (history.getMemberId() == null || !history.getMemberId().equals(memberId)) {
+            throw new AccessDeniedException(ErrorCode.ACCESS_DENIED);
         }
-    }
 
-    private Integer parseAmount(String amountStr) {
-        if (amountStr == null || amountStr.isBlank()) {
-            return null;
-        }
-        try {
-            String numberOnly = amountStr.replaceAll("[^0-9]", "");
-            return numberOnly.isEmpty() ? null : Integer.parseInt(numberOnly);
-        } catch (NumberFormatException e) {
-            return null;
-        }
-    }
-
-    @Transactional
-    private void saveSteps(Recipe recipe, List<StepParseResult> steps) {
-        for (StepParseResult step : steps) {
-            RecipeStep recipeStep = RecipeStep.create(
-                    recipe, step.stepNumber(), step.description());
-            recipeStepRepository.save(recipeStep);
-        }
-    }
-
-    private void addToDefaultRecipeBook(Long memberId, Long recipeId) {
-        recipeBookQueryService.findAllByMemberId(memberId).stream()
-                .filter(RecipeBook::getIsDefault)
-                .findFirst()
-                .ifPresent(defaultBook -> recipeBookService.addRecipeInternal(
-                        memberId, defaultBook.getId(), recipeId));
+        return RecipeImportStatusResponse.from(history);
     }
 
     private Member findMemberById(Long memberId) {
         return memberRepository.findById(memberId)
                 .orElseThrow(() -> new NotFoundException(ErrorCode.MEMBER_NOT_FOUND));
-    }
-
-    private CuisineType parseCuisineType(String value) {
-        if (value == null)
-            return null;
-        try {
-            return CuisineType.valueOf(value);
-        } catch (IllegalArgumentException e) {
-            log.warn("Unknown cuisineType: {}", value);
-            return null;
-        }
-    }
-
-    private MealType parseMealType(String value) {
-        if (value == null)
-            return null;
-        try {
-            return MealType.valueOf(value);
-        } catch (IllegalArgumentException e) {
-            log.warn("Unknown mealType: {}", value);
-            return null;
-        }
-    }
-
-    private Difficulty parseDifficulty(String value) {
-        if (value == null)
-            return null;
-        try {
-            return Difficulty.valueOf(value);
-        } catch (IllegalArgumentException e) {
-            log.warn("Unknown difficulty: {}", value);
-            return null;
-        }
     }
 
     private void validateNotDuplicateSource(String sourceUrl) {
