+++ src/main/java/com/shortkki/global/config/AsyncConfig.java
@@ -1,0 +1,9 @@
+package com.shortkki.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableAsync;
+
+@Configuration
+@EnableAsync
+public class AsyncConfig {
+}
